pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "programacaiii_api"
version = "0.1.0"
description = "API para a disciplina de Programação III"
authors = ["Murilo <murilo.assis@ifg.edu.br.com>"]
readme = "README.md"
packages = [
    { include = "main.py" },
    { include = "app" }
]

[tool.poetry.dependencies]
python = "3.11.*"

# Dependências da sua API (usando as versões do [project])
fastapi = ">=0.116.1,<0.117.0"
uvicorn = ">=0.35.0,<0.36.0"
email-validator = ">=2.3.0,<3.0.0"
sqlalchemy = ">=2.0.43,<3.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
python-jose = ">=3.5.0,<4.0.0"
python-multipart = ">=0.0.20,<1.0.0" # Corrigido do conflito
passlib = {extras = ["bcrypt"], version = ">=1.7.4,<2.0.0"}
bcrypt = ">=4.0.1,<5.0.0"
pillow = ">=11.3.0,<12.0.0"
web3 = ">=6.3.0,<7.0.0"

# Dependências de Teste (que só estavam em [tool.poetry.dependencies])
pytest = "^8.4.2"
httpx = "^0.28.1"


[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", "app"]